initial commit
unit test does not pass, returns a list of DailyPortFolio objects with tthe correct dates and in the  correct order but does not calculate value correctly for dates that are not in the PRICES list.

Code attempted earlier has been left commented out, currently all code is contained in the one method with no separation of concerns and would benefit from refactoring.